+++ src/types/rx-collection.d.ts
@@ -8,6 +8,7 @@ import type {
     RxCollectionBase
 } from '../rx-collection';
 import type { QueryCache } from '../query-cache';
+import { Observable } from 'rxjs';
 
 export interface KeyFunctionMap {
     [key: string]: Function;
@@ -83,6 +84,7 @@ export interface RxCollectionGenerated<RxDocumentType = any, OrmMethods = {}> {
     insertLocal(id: string, data: any): Promise<RxLocalDocument<RxCollection<RxDocumentType, OrmMethods>>>;
     upsertLocal<LocalDocType = any>(id: string, data: LocalDocType): Promise<RxLocalDocument<RxCollection<RxDocumentType, OrmMethods>, LocalDocType>>;
     getLocal<LocalDocType = any>(id: string): Promise<RxLocalDocument<RxCollection<RxDocumentType, OrmMethods>, LocalDocType>>;
+    getLocal$<LocalDocType = any>(id: string): Observable<RxLocalDocument<RxCollection<RxDocumentType, OrmMethods>, LocalDocType> | null>;
 
     // only inMemory-collections
     awaitPersistence(): Promise<void>;