+++ test/helper/humans-collection.ts
@@ -19,7 +19,7 @@ export async function create(
     size: number = 20,
     name: string = 'human',
     multiInstance: boolean = true
-): Promise<RxCollection<HumanDocumentType>> {
+): Promise<RxCollection<HumanDocumentType, {}, {}>> {
     if (!name) name = 'human';
     PouchDB.plugin(require('pouchdb-adapter-memory'));
     const db = await createRxDatabase<{ human: RxCollection<schemaObjects.HumanDocumentType> }>({
@@ -48,7 +48,7 @@ export async function create(
 export async function createBySchema<RxDocumentType = {}>(
     schema: RxJsonSchema,
     name = 'human'
-): Promise<RxCollection<RxDocumentType>> {
+): Promise<RxCollection<RxDocumentType, {}, {}>> {
     PouchDB.plugin(require('pouchdb-adapter-memory'));
 
     const db = await createRxDatabase<{ [prop: string]: RxCollection<RxDocumentType> }>({
@@ -71,7 +71,7 @@ export async function createAttachments(
     size = 20,
     name = 'human',
     multiInstance = true
-): Promise<RxCollection<schemaObjects.HumanDocumentType>> {
+): Promise<RxCollection<schemaObjects.HumanDocumentType, {}, {}>> {
 
     if (!name) name = 'human';
     PouchDB.plugin(require('pouchdb-adapter-memory'));
@@ -106,7 +106,7 @@ export async function createEncryptedAttachments(
     size = 20,
     name = 'human',
     multiInstance = true
-): Promise<RxCollection<schemaObjects.HumanDocumentType>> {
+): Promise<RxCollection<schemaObjects.HumanDocumentType, {}, {}>> {
 
     if (!name) name = 'human';
     PouchDB.plugin(require('pouchdb-adapter-memory'));
@@ -340,7 +340,7 @@ export async function createMultiInstance(
     name: string,
     amount = 0,
     password = null
-): Promise<RxCollection<schemaObjects.HumanDocumentType>> {
+): Promise<RxCollection<schemaObjects.HumanDocumentType, {}, {}>> {
     PouchDB.plugin(require('pouchdb-adapter-memory'));
 
     const db = await createRxDatabase<{ human: RxCollection<schemaObjects.HumanDocumentType> }>({